+++ app/src/main/AndroidManifest.xml
@@ -18,5 +18,9 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <service
+        android:name=".background.Service"
+        android:exported="false"/>
+
     </application>
 </manifest>
